Created detector module, renamed others back to _Io.py
