+++ src/dodal/devices/electron_analyser/__init__.py
@@ -1,8 +1,9 @@
 from .abstract_region import EnergyMode
-from .specs_analyser import SpecsAnalyserDetector, SpecsAnalyserDriverIO
+from .detector import ElectronAnalyserDetector, ElectronAnalyserRegionDetector
+from .specs_analyser_io import SpecsAnalyserDriverIO
 from .specs_region import SpecsRegion, SpecsSequence
 from .util import to_binding_energy, to_kinetic_energy
-from .vgscienta_analyser import VGScientaAnalyserDetector, VGScientaAnalyserDriverIO
+from .vgscienta_analyser_io import VGScientaAnalyserDriverIO
 from .vgscienta_region import (
     VGScientaExcitationEnergySource,
     VGScientaRegion,
@@ -11,12 +12,12 @@ from .vgscienta_region import (
 
 __all__ = [
     "EnergyMode",
+    "ElectronAnalyserDetector",
+    "ElectronAnalyserRegionDetector",
     "SpecsAnalyserDriverIO",
-    "SpecsAnalyserDetector",
     "SpecsRegion",
     "SpecsSequence",
     "VGScientaAnalyserDriverIO",
-    "VGScientaAnalyserDetector",
     "VGScientaExcitationEnergySource",
     "VGScientaRegion",
     "VGScientaSequence",
