+++ src/dodal/devices/electron_analyser/vgscienta_analyser_io.py
@@ -2,11 +2,8 @@ import numpy as np
 from ophyd_async.core import Array1D, SignalR, StandardReadableFormat, soft_signal_rw
 from ophyd_async.epics.core import epics_signal_r, epics_signal_rw
 
-from dodal.common.beamlines.device_helpers import CAM_SUFFIX
-from dodal.common.data_util import load_json_file_to_class
-from dodal.devices.electron_analyser.abstract_analyser import (
+from dodal.devices.electron_analyser.abstract_analyser_io import (
     AbstractAnalyserDriverIO,
-    AbstractElectronAnalyserDetector,
 )
 from dodal.devices.electron_analyser.vgscienta_region import (
     DetectorMode,
@@ -14,7 +11,7 @@ from dodal.devices.electron_analyser.vgscienta_region import (
 )
 
 
-class VGScientaAnalyserDriverIO(AbstractAnalyserDriverIO):
+class VGScientaAnalyserDriverIO(AbstractAnalyserDriverIO[VGScientaSequence]):
     def __init__(self, prefix: str, name: str = "") -> None:
         with self.add_children_as_readables(StandardReadableFormat.CONFIG_SIGNAL):
             self.excitation_energy_source = soft_signal_rw(str, initial_value=None)
@@ -42,15 +39,5 @@ class VGScientaAnalyserDriverIO(AbstractAnalyserDriverIO):
     def pass_energy_type(self) -> type:
         return str
 
-
-class VGScientaAnalyserDetector(
-    AbstractElectronAnalyserDetector[VGScientaAnalyserDriverIO, VGScientaSequence]
-):
-    def __init__(self, prefix: str, name: str = ""):
-        self.driver = VGScientaAnalyserDriverIO(
-            prefix=prefix + CAM_SUFFIX, name="driver"
-        )
-        super().__init__(name, self.driver)
-
-    def get_sequence(self, filename: str) -> VGScientaSequence:
-        return load_json_file_to_class(VGScientaSequence, filename)
+    def sequence_type(self) -> type[VGScientaSequence]:
+        return VGScientaSequence
